% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaf_generative_model.R
\name{add_tail_muts}
\alias{add_tail_muts}
\title{Associate number of tail mutations to the leaves of a sample tree.}
\usage{
add_tail_muts(tree, mu, len = 2.7 * 10^9, vaf_min = 0.05)
}
\arguments{
\item{tree}{A sample tree}

\item{mu}{Mutation rate}

\item{len}{Length of the genome}

\item{vaf_min}{Minimum detectable VAF of a mutation}
}
\value{
a table of mutations with number of variants and depth
}
\description{
Number of tail mutations is assigned to any leave based on its ccf and mutation rate
}
\examples{
add_tail_muts(tree,1e-07,2.7*10^9,0.05)
}
