% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaf_generative_model.R
\name{calculate_ccf}
\alias{calculate_ccf}
\title{Calculate ccf for nodes of a sample tree.}
\usage{
calculate_ccf(tree, s)
}
\arguments{
\item{tree}{A sample tree}

\item{s}{Fitness advantage of + cells with respect to - cells}
}
\value{
sample tree
}
\description{
CCFs are computed according to number of mutations assigned to any node and fitness advantage.
}
\examples{
calculate_ccf(tree,s)
}
