% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_input_count}
\alias{check_input_count}
\title{Check counts input}
\usage{
check_input_count(counts)
}
\arguments{
\item{counts}{A tibble or data.frame with columns `count_n`, `count_p`, `time`}
}
\value{
Invisibly returns TRUE if checks pass; stops with error otherwise.
}
\description{
Check counts input
}
