% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_input_counts}
\alias{check_input_counts}
\title{Check if counts input data has the correct form.
Check if the input data has the correct fields and stop if someone is missing.}
\usage{
check_input_counts(data)
}
\arguments{
\item{data}{Dataframe containing with number of cell counts of - and + cells 
at different time points in case of Counts application.}
}
\value{
PEPI object
}
\description{
Check if counts input data has the correct form.
Check if the input data has the correct fields and stop if someone is missing.
}
\examples{
check_input_counts(data)
}
