% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_input_vaf}
\alias{check_input_vaf}
\title{Check clade_statistics input for arbitrary time points}
\usage{
check_input_vaf(clade_statistics)
}
\arguments{
\item{clade_statistics}{A tibble or data.frame with clade-level statistics.
Must include columns `cluster_name`, `n_muts`, `is_clade`, `names`, `driver_type`, `has_driver`, `rewind`
and VAF columns named `vaf_k_n` and `vaf_k_p` for k = 1, ..., N.}
}
\value{
Invisibly returns TRUE if checks pass; stops with error otherwise.
}
\description{
Check clade_statistics input for arbitrary time points
}
