% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_generators.R
\name{counts_inference_code}
\alias{counts_inference_code}
\title{Generate stan model for counts inference.}
\usage{
counts_inference_code(likelihood = T)
}
\arguments{
\item{likelihood}{Boolean value specifying to include the likelihood in the model, or just sampling from the priors.}
}
\value{
A string containing a stan model
}
\description{
A stan file with the counts inference model is generated.
}
\examples{
counts_inference_code(likelihood = T)
}
