% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{driver_example}
\alias{driver_example}
\title{Dataset for driver analysis}
\format{
A data frame or list with example data
}
\usage{
data(driver_example)
}
\description{
Dataset for driver analysis
}
\keyword{datasets}
