% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaf_generative_model.R
\name{filter_nodes}
\alias{filter_nodes}
\title{Prunes a sample tree.}
\usage{
filter_nodes(tree, threshold = 0.01)
}
\arguments{
\item{tree}{A sample tree}

\item{threshold}{Pruning threshold for epimutation probability}
}
\value{
sample tree
}
\description{
Nodes with low epimutation probability are removed and replaced with a leave.
}
\examples{
filter_nodes(tree,threshold = 0.01)
}
