% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitters.R
\name{fit_counts}
\alias{fit_counts}
\title{A PEPI fit with cell counts inference is returned.}
\usage{
fit_counts(
  x,
  path_to_model,
  cmdstan_path,
  ndraws = 1000,
  init = NULL,
  seed = 45,
  alpha_ln = 1.5,
  beta_ln = 1,
  alpha_lp = 1.5,
  beta_lp = 1,
  alpha_rn = 1,
  beta_rn = 10,
  alpha_rp = 1,
  beta_rp = 10
)
}
\arguments{
\item{x}{PEPI object containing cell counts data}

\item{ndraws}{Number of draws from the posterior}

\item{init}{List of initialization parameters}

\item{seed}{Seed of the computation}

\item{alpha_ln}{shape parameter for gamma prior on - growth rate}

\item{beta_ln}{rate parameter for gamma prior on - growth rate}

\item{alpha_lp}{shape parameter for gamma prior on + growth rate}

\item{beta_lp}{rate parameter for gamma prior on + growth rate}

\item{alpha_rn}{shape parameter for gamma prior on - effective switch rate}

\item{beta_rn}{rate parameter for gamma prior on - effective switch rate}

\item{alpha_rp}{shape parameter for gamma prior on + effective switch rate}

\item{beta_rp}{rate parameter for gamma prior on + effective switch rate}

\item{threshold}{Threshold for tree pruning}
}
\value{
PEPI fit
}
\description{
A PEPI fit with cell counts inference is returned.
}
\examples{
fit_counts(x,path_to_model,cmdstan_path,
                     ndraws = 1000,init = NULL,seed = 45, alpha_ln = 1.5,
                     beta_ln = 1, alpha_lp = 1.5, beta_lp = 1, alpha_rn = 1, beta_rn = 10,
                     alpha_rp = 1, beta_rp = 10)
}
