% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitters.R
\name{fit_s}
\alias{fit_s}
\title{Infer epimutation clocks in number of cell divisions and fitness of + cells with respect to - cells.}
\usage{
fit_s(
  x,
  path_to_model,
  cmdstan_path,
  threshold = 0.1,
  ndraws = 1000,
  init = NULL,
  seed = 45,
  mu = 1e-07,
  l = 2.7 * 10^9,
  ms = -0.5,
  sigma = 0.5,
  k = 100
)
}
\arguments{
\item{x}{PEPI object containing fitness and epimutation clocks inference}

\item{path_to_model}{String specifying the path where we want to save the stan model for a given depth}

\item{cmdstan_path}{String specifying the path to cmdstan folder}

\item{threshold}{Threshold for tree pruning}

\item{ndraws}{Number of draws from the posterior}

\item{init}{List of initialization parameters}

\item{seed}{Seed of the computation}

\item{mu}{Mutation rate per division per bp per allele}

\item{l}{length of the genome}

\item{ms}{Mean of lognormal prior for s
@param sigma Sigma parameter of lognormal prior for s}

\item{k}{Number of trials for the beta prior on cluster centroids}
}
\value{
PEPI object
}
\description{
A PEPI fit with fitness inference is returned.
}
\examples{
fit_s(x,path_to_model = "models",cmdstan_path = "my_cmdstan/",threshold = 0.1,
ndraws = 1000,init = NULL,seed = 45,
mu = 1e-7,l = 2.7*10^9,ms = -0.5,sigma = 0.5,k = 100)
}
