% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitters.R
\name{fit_tree}
\alias{fit_tree}
\title{Fit a multivariate vaf spectrum with epigenetic tree model.}
\usage{
fit_tree(
  x,
  path_to_model,
  cmdstan_path,
  max_depth = 2,
  ndraws = 1000,
  init = NULL,
  seed = 15,
  mu = 1e-07,
  l = 2.7 * 10^9,
  rho_n = 1,
  rho_p = 1,
  nu_t = 0.1,
  qt = 10000,
  rate_n = 0.001,
  qn = 10000,
  rate_p = 0.001,
  qp = 10000,
  k = 10000,
  gamma = 150
)
}
\arguments{
\item{x}{PEPI object containing VAF multivariate spectrum}

\item{path_to_model}{String specifying the path where we wannt to save the stan model for a given depth}

\item{cmdstan_path}{String specifying the path to cmdstan folder}

\item{max_depth}{Maximum number of levels}

\item{ndraws}{Number of draws from the posterior}

\item{init}{List of initialization parameters}

\item{seed}{Seed of the computation}

\item{mu}{Mutation rate per division per bp per allele}

\item{l}{length of the genome}

\item{rho_n}{Purity of - sample
@param rho_p Purity of + sample}

\item{nu_t}{Mean of the beta prior on fraction of truncal mutations}

\item{qt}{Number of trials of the beta prior on fraction of truncal mutations}

\item{rate_n}{Mean of the beta prior on the epimutation rate from - to +}

\item{qn}{Number of trials for the beta prior on the epimutation rate from - to +}

\item{rate_p}{Mean of the beta prior on the epimutation rate from + to -}

\item{qp}{Number of trials for the beta prior on the epimutation rate from - to +}

\item{k}{Number of trials for the beta prior on cluster centroids}

\item{gamma}{Concentration of a Dirichlet distribution to split mutations at any node}
}
\value{
PEPI object
}
\description{
A PEPI fit with tree inference is returned.
}
\examples{
fit_tree(x,path_to_model = "models",cmdstan_path = "my_cmdstan/",
max_depth = 2,ndraws = 1000,init = NULL,seed = 15,
mu = 1e-7,l = 2.7*10^9,rho_n = 1,rho_p = 1,nu_t = 0.1,
qt = 1e4,rate_n = 1e-3,qn = 1e4,rate_p = 1e-3,
qp = 1e4,k = 1e4,gamma = 150)
}
