% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_generators.R
\name{fitness_inference_code}
\alias{fitness_inference_code}
\title{Generate stan model for fitness inference.}
\usage{
fitness_inference_code(tree, likelihood = T)
}
\arguments{
\item{tree}{Sample tree with epigenetic state associated to any node}

\item{likelihood}{Boolean value specifying to include the likelihood in the model, or just sampling from the priors.}
}
\value{
A string containing a stan model
}
\description{
A stan file with the fitness inference model is generated.
}
\examples{
fitness_inference_code(tree,likelihood = T)
}
