% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaf_generative_model.R
\name{generate_nodes}
\alias{generate_nodes}
\title{generative models
Generate a sample tree with input epimutation rates, fraction of truncal mutations and fitness advantage between + and -.}
\usage{
generate_nodes(
  M,
  gamma,
  at,
  bt,
  s,
  rate_minus,
  rate_plus,
  mu,
  len,
  max_depth = 3
)
}
\arguments{
\item{M}{total number of mutations}

\item{gamma}{parameter of a beta distribution controlling how we split mutations at any node}

\item{at}{number of successes of a beta distribution controlling amount of truncal mutations}

\item{bt}{number of fails of a beta distribution controlling amount of truncal mutations}

\item{s}{Fitness advantage of + cells with respect to - cells}

\item{rate_minus}{Average number of epimutation from - to + per mutations}

\item{rate_plus}{Average number of epimutation from + to - per mutations}

\item{mu}{Mutation rate per division per bp per allele}

\item{len}{length of the genome}

\item{max_depth}{maximal level of the tree}
}
\value{
sample tree
}
\description{
A tree with number of mutations per nodes is generated.
}
\examples{
generate_nodes(6000,100,10,80,0.2,1e-3,1e-3,1e-7,2.7*10^9,3)
}
