% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaf_generative_model.R
\name{generate_spectrum}
\alias{generate_spectrum}
\title{Simulate VAF multivariate spectrum associated to a sample tree.}
\usage{
generate_spectrum(tree, vaf_min, DP, tail = T)
}
\arguments{
\item{tree}{A sample tree}

\item{vaf_min}{Lower bound on VAF spectrum}

\item{DP}{average coverage}
}
\value{
a table of mutations with number of variants and depth
}
\description{
A bivariate binomial process is simulated any node, with number of trials and success probabilities given by number of mutations and ccfs. If tail mutations
are included, thei VAF is sampled from a power law distribution.
}
\examples{
generate_spectrum(tree,vaf_min = 0.05,DP = 150,tail = T)
}
