% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_average_counts}
\alias{get_average_counts}
\title{Get average predicted cell counts from the fit.}
\usage{
get_average_counts(x)
}
\arguments{
\item{x}{PEPI object}
}
\value{
A PEPI object with average counts
}
\description{
Posterior draws are extracted from the fit.
}
\examples{
get_average_counts(x)
}
