% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_average_tree}
\alias{get_average_tree}
\title{utilities inference
Get the inferred tree from a fit.}
\usage{
get_average_tree(fit, threshold = NULL)
}
\arguments{
\item{threshold}{Threshold on epimutation probability for tree pruning}

\item{x}{A PEPI fit}
}
\value{
PEPI fit with inferred average sample tree
}
\description{
A tree is obtained from the inferred values of nodes ccfs and number of mutations.
}
\examples{
get_average_tree(x,threshold = 0.1)
}
