% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_clusters}
\alias{get_clusters}
\title{Infer clusters from data and inferred tree.}
\usage{
get_clusters(fit)
}
\arguments{
\item{fit}{PEPI object}
}
\value{
A PEPI object containing membership probabilities and cluster assignents
}
\description{
Membership probabilities and cluster assignents are computed for any mutation.
}
\examples{
get_clusters(fit)
}
