% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_init_values}
\alias{get_init_values}
\title{Provide a initialization parameters for a PEPI VAF fit}
\usage{
get_init_values(spectrum, K = 10, alpha = 10, samples = 1, pi_cutoff = 0.01)
}
\arguments{
\item{spectrum}{Mutation data with number of variants and depth for any mutation and sample}

\item{K}{Maximum number of clusters}

\item{alpha}{Dirichlet concentration parameter}

\item{samples}{Number of fits computed by the algorithm}

\item{pi_cutoff}{Cutoff on mixing proportions to filter clusters and reassign mutations}
}
\value{
a list with initialization values
}
\description{
Return a list of initialization values for some parameters through some euristics.
}
\examples{
\dontrun{
get_init_values(spectrum,K = 10,alpha = 10,samples = 1, pi_cutoff = 0.01)
}
}
