% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_n_times}
\alias{get_n_times}
\title{Infer number of time points from a PEPI object}
\usage{
get_n_times(pepi)
}
\arguments{
\item{pepi}{A PEPI object.}
}
\value{
Integer number of time points.
}
\description{
Infer number of time points from a PEPI object
}
