% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_posterior}
\alias{get_posterior}
\title{Extract and save posterior in a PEPI object}
\usage{
get_posterior(pepi)
}
\arguments{
\item{pepi}{A PEPI object containing a Stan fit (`fit`).}
}
\value{
The updated PEPI object with posterior saved in pepi$posterior.
}
\description{
Extract and save posterior in a PEPI object
}
