% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_predicted_ccf}
\alias{get_predicted_ccf}
\title{Get posterior-predictive CCF (ccf_*_pred) with observed CCF joined in.}
\usage{
get_predicted_ccf(x)
}
\arguments{
\item{x}{PEPI_Multirates object, after \code{get_posterior_multirates()}.}
}
\value{
A tibble with posterior mean/95%CI and the matching observed CCF.
}
\description{
Get posterior-predictive CCF (ccf_*_pred) with observed CCF joined in.
}
\examples{
\dontrun{
get_predicted_ccf(x)
}
}
