% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_predicted_counts}
\alias{get_predicted_counts}
\title{Get posterior-predictive counts (z_*_pred) with observed counts joined in.}
\usage{
get_predicted_counts(x)
}
\arguments{
\item{x}{PEPI_Multirates object, after \code{get_posterior_multirates()}.}
}
\value{
A tibble with posterior mean/95%CI and the matching observed count.
}
\description{
Get posterior-predictive counts (z_*_pred) with observed counts joined in.
}
\examples{
\dontrun{
get_predicted_counts(x)
}
}
