% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_predicted_fractions}
\alias{get_predicted_fractions}
\title{Get posterior-predictive fractions (frac_*_pred) with observed fractions joined in.}
\usage{
get_predicted_fractions(x)
}
\arguments{
\item{x}{PEPI_Multirates object, after \code{get_posterior_multirates()}.}
}
\value{
A tibble with posterior mean/95%CI and the matching observed fraction.
}
\description{
Get posterior-predictive fractions (frac_*_pred) with observed fractions joined in.
}
\examples{
\dontrun{
get_predicted_fractions(x)
}
}
