% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_prior}
\alias{get_prior}
\title{to fix
Get draws from prior used for the inference}
\usage{
get_prior(x, model_types = c("tree", "counts"), ndraws = 1000)
}
\arguments{
\item{x}{PEPI object}

\item{model_types}{list of models we want the prior}

\item{ndraws}{Number of draws}
}
\value{
A PEPI object with prior draws
}
\description{
Prior draws are generated from PEPI object.
}
\examples{
get_prior(x,model_type = c("tree","counts"),ndraws = 1000)
}
