% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_stan_data_pepi}
\alias{get_stan_data_pepi}
\title{Build Stan data list for PEPI model}
\usage{
get_stan_data_pepi(pepi)
}
\arguments{
\item{pepi}{A PEPI object.}
}
\value{
Named list for Stan input.
}
\description{
Build Stan data list for PEPI model
}
