% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inits.R
\name{init_counts}
\alias{init_counts}
\title{Create a PEPI object of Counts type.}
\usage{
init_counts(data)
}
\arguments{
\item{data}{Dataframe containing with number of cell counts of - and + cells at different time points}
}
\value{
PEPI object
}
\description{
A PEPI object of vaf type or counts type is created.
}
\examples{
init_counts(data)
}
