% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inits.R
\name{init_vaf}
\alias{init_vaf}
\title{Create a PEPI object of VAF type.}
\usage{
init_vaf(data)
}
\arguments{
\item{data}{Dataset to analyze. It must be a dataframe with number of variants and depth for any mutation and sample}
}
\value{
PEPI object
}
\description{
A PEPI object vaf type is created.
}
\examples{
init_vaf(data)
}
