% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{no_driver_example}
\alias{no_driver_example}
\title{Dataset for no_driver analysis}
\format{
A data frame or list with example data
}
\usage{
data(no_driver_example)
}
\description{
Dataset for no_driver analysis
}
\keyword{datasets}
