% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_acf}
\alias{plot_acf}
\title{Autocorrelation of the sampled chains.}
\usage{
plot_acf(x, params = NULL, lags = 20)
}
\arguments{
\item{x}{PEPI_Multirates object fit with \code{method = "sample"}.}

\item{params}{Vector of Stan parameter base names. Defaults to a
handful of global scalar parameters.}

\item{lags}{Number of lags to show.}
}
\value{
A ggplot object (see \code{bayesplot::mcmc_acf()}).
}
\description{
Autocorrelation of the sampled chains.
}
\examples{
\dontrun{
plot_acf(x)
}
}
