% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_ccf}
\alias{plot_ccf}
\title{Plot predicted vs observed CCF for the unified multirates model.}
\usage{
plot_ccf(x)
}
\arguments{
\item{x}{PEPI_Multirates object, after \code{get_posterior_multirates()}.}
}
\value{
Plot of predicted vs observed CCF.
}
\description{
Faceted by group/id and epistate, using \code{get_predicted_ccf()}.
}
\examples{
\dontrun{
plot_ccf(x)
}
}
