% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_counts}
\alias{plot_counts}
\title{Plot predicted vs input counts.}
\usage{
plot_counts(x)
}
\arguments{
\item{x}{PEPI object}
}
\value{
Plot of the counts.
}
\description{
A plot of predicted and input counts is generated.
}
\examples{
plot_counts(x)
}
