% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_counts_multirates}
\alias{plot_counts_multirates}
\title{Plot predicted vs observed counts for the unified multirates model.}
\usage{
plot_counts_multirates(x)
}
\arguments{
\item{x}{PEPI_Multirates object, after \code{get_posterior_multirates()}.}
}
\value{
Plot of predicted vs observed counts.
}
\description{
Faceted by group/id and epistate, using \code{get_predicted_counts()}.
}
\examples{
\dontrun{
plot_counts_multirates(x)
}
}
