% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_divergences}
\alias{plot_divergences}
\title{Divergent transitions.}
\usage{
plot_divergences(x)
}
\arguments{
\item{x}{PEPI_Multirates object fit with \code{method = "sample"}.}
}
\value{
A ggplot object (see \code{bayesplot::mcmc_nuts_divergence()}).
}
\description{
Divergences concentrating in a region of parameter space indicate the
sampler struggled there and posterior estimates may be biased.
}
\examples{
\dontrun{
plot_divergences(x)
}
}
