% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_forest}
\alias{plot_forest}
\title{Plot a forest of cells}
\usage{
plot_forest(forest)
}
\arguments{
\item{forest}{A data.frame representing the forest. Must contain columns:
\code{cell_id}, \code{ancestor}, \code{mutant}, \code{epistate}, \code{sample}, \code{birth_time}.}
}
\value{
A ggplot2 object representing the forest as a tree.
}
\description{
Plot a forest of cells
}
