% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_marginal}
\alias{plot_marginal}
\title{Plot marginal VAF distributions with cluster associated to tree nodes.}
\usage{
plot_marginal(spectrum)
}
\arguments{
\item{spectrum}{VAF spectrum with cluster labels}
}
\value{
Two marginal histograms.
}
\description{
Two marginal histograms are generated from a labelled dataset.
}
\examples{
\dontrun{
plot_marginal(spectrum)
}
}
