% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_posterior_predictive}
\alias{plot_posterior_predictive}
\title{Plot posterior predictive for a parameter}
\usage{
plot_posterior_predictive(pepi, parameter, log_scale = TRUE)
}
\arguments{
\item{pepi}{A PEPI object with `posterior` extracted.}

\item{parameter}{Name of the posterior predictive variable (starts with "pred").}

\item{log_scale}{Logical, whether to plot x-axis on log scale.}
}
\value{
A ggplot object.
}
\description{
Plot posterior predictive for a parameter
}
