% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot sample tree with branches associated to nodes.}
\usage{
plot_tree(tree)
}
\arguments{
\item{x}{Tree object}
}
\value{
Plot of the tree.
}
\description{
A ggtree plot of the sample tree is generated from the inferred tree.
}
\examples{
plot_tree(tree)
}
