% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_vaf}
\alias{plot_vaf}
\title{Plot VAF distributions and pairwise scatterplots}
\usage{
plot_vaf(muts, variables = NULL, vaf_cut = 0.02)
}
\arguments{
\item{muts}{A data.frame containing mutation VAFs. Column names must be like vaf_1_n, vaf_1_p, etc.}

\item{variables}{Character vector of columns to plot. Must exist in muts.}

\item{vaf_cut}{Minimum VAF to include in plots (default 0.02).}
}
\value{
A list of ggplot objects: marginals and, if more than 1 variable, pairwise scatterplots.
}
\description{
Plot VAF distributions and pairwise scatterplots
}
