% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rewind_example}
\alias{rewind_example}
\title{Dataset for rewind analysis}
\format{
A data frame or list with example data
}
\usage{
data(rewind_example)
}
\description{
Dataset for rewind analysis
}
\keyword{datasets}
