% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_generators.R
\name{tree_inference_code}
\alias{tree_inference_code}
\title{Generate stan model for tree inference.}
\usage{
tree_inference_code(max_depth, likelihood = T)
}
\arguments{
\item{max_depth}{Depth of the tree}

\item{likelihood}{Boolean value specifying to include the likelihood in the model, or just sampling from the priors.}
}
\value{
A string containing a stan model
}
\description{
A stan file with the tree inference model is generated.
}
\examples{
tree_inference_code(max_depth = 2,likelihood = T)
}
