% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_input_vaf}
\alias{check_input_vaf}
\title{Check if VAF input data has the correct form.
Check if the input data has the correct fields and stop if someone is missing.}
\usage{
check_input_vaf(data)
}
\arguments{
\item{data}{Dataset to analyze. It must be a dataframe with number of variants and depth 
for any mutation and sample.}
}
\value{
PEPI object
}
\description{
Check if VAF input data has the correct form.
Check if the input data has the correct fields and stop if someone is missing.
}
\examples{
check_input_vaf(data)
}
