% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_colors}
\alias{get_colors}
\title{Associate colors to nodes.}
\usage{
get_colors(max_depth)
}
\arguments{
\item{max_depth}{Maximal number of levels of the tree}
}
\value{
Named list of colors.
}
\description{
A list of colors labelled by nodes is generated.
}
\examples{
get_colors(max_depth = 2)
}
