% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_pepi_theme}
\alias{get_pepi_theme}
\title{Generate personalized ggplot theme.}
\usage{
get_pepi_theme()
}
\value{
a ggplot theme.
}
\description{
A ggplot theme is generated.
}
\examples{
PEPI:::get_pepi_theme()
}
\keyword{internal}
