% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_posterior}
\alias{get_posterior}
\title{Get posterior distribution from PEPI object.}
\usage{
get_posterior(fit)
}
\arguments{
\item{x}{PEPI object}
}
\value{
A PEPI object with posterior draws
}
\description{
Posterior draws are extracted from the fit.
}
\examples{
get_posterior(x)
}
