% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_prior}
\alias{get_prior}
\title{Extract and save prior in a PEPI object}
\usage{
get_prior(pepi)
}
\arguments{
\item{pepi}{A PEPI object containing a Stan fit (`fit`).}
}
\value{
The updated PEPI object with prior saved in pepi$prior.
}
\description{
Extract and save prior in a PEPI object
}
