% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_viber_clusters}
\alias{get_viber_clusters}
\title{Clusters mutation with VIBER, a package that implements a variational
non parametric Bayesian model to fit multi-variate Binomial mixtures}
\usage{
get_viber_clusters(data, K = 10, alpha = 10, samples = 1, pi_cutoff = 0.01)
}
\arguments{
\item{data}{Mutation data with number of variants and depth for any mutation and sample}

\item{K}{Maximum number of clusters}

\item{alpha}{Dirichlet concentration parameter}

\item{samples}{Number of fits computed by the algorithm}

\item{pi_cutoff}{Cutoff on mixing proportions to filter clusters and reassign mutations}
}
\value{
a tibble with clusters names, mixing proportion and vaf coordinates
}
\description{
Clusters with centroids VAF coordiates and relative proportion are given.
}
\examples{
\dontrun{
# requires the VIBER package (github.com/caravagn/VIBER), not on CRAN
set.seed(1)
data = data.frame(
  Nx = rbinom(50, 100, 0.3), DPx = rep(100, 50),
  Ny = rbinom(50, 100, 0.1), DPy = rep(100, 50)
)
get_viber_clusters(data, K = 10, alpha = 1, samples = 1, pi_cutoff = 0.01)
}
}
