% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_fractions}
\alias{plot_fractions}
\title{Plot predicted vs observed fractions for the unified multirates model.}
\usage{
plot_fractions(x)
}
\arguments{
\item{x}{PEPI_Multirates object, after \code{get_posterior_multirates()}.}
}
\value{
Plot of predicted vs observed fractions.
}
\description{
Faceted by group/id and epistate, using \code{get_predicted_fractions()}.
}
\examples{
\dontrun{
plot_fractions(x)
}
}
