% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_inference}
\alias{plot_inference}
\title{Plot prior vs posterior for a parameter in a PEPI object}
\usage{
plot_inference(pepi, parameter, log_scale = FALSE)
}
\arguments{
\item{pepi}{A PEPI object with `posterior` and `prior` extracted.}

\item{parameter}{Name of the parameter to plot.}

\item{log_scale}{Logical, whether to plot x-axis on log scale.}
}
\value{
A ggplot object.
}
\description{
Plot prior vs posterior for a parameter in a PEPI object
}
