% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_inference}
\alias{plot_inference}
\title{Plot posterior and prior distributions.}
\usage{
plot_inference(x, params = NULL)
}
\arguments{
\item{params}{A list of parameters}

\item{post}{Posterior draws}
}
\value{
A plot with posterior and prior distributions
}
\description{
Posterior and prior draws histograms are plotted for any required parameter.
}
\examples{
plot_inference(x,params = c("rn","rp"))
}
