% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_marginal}
\alias{plot_marginal}
\title{Plot marginal VAF distributions with cluster associated to tree nodes.}
\usage{
plot_marginal(spectrum)
}
\arguments{
\item{spectrum}{VAF spectrum with cluster labels}
}
\value{
Two marginal histograms.
}
\description{
Two marginal histograms are generated from a labelled dataset.
}
\examples{
library(dplyr)
library(ggplot2)
library(ggpubr)
set.seed(1)
spectrum = data.frame(
  Nx = rbinom(50, 100, 0.3), DPx = rep(100, 50),
  Ny = rbinom(50, 100, 0.1), DPy = rep(100, 50),
  node = sample(c("-","+"), 50, replace = TRUE)
)
plot_marginal(spectrum)
}
