% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_multivariate}
\alias{plot_multivariate}
\title{plotting functions
Plot multivariate VAF distributions with cluster associated to tree nodes.}
\usage{
plot_multivariate(spectrum)
}
\arguments{
\item{spectrum}{VAF spectrum with cluster labels}
}
\value{
A multivariate plot
}
\description{
A multivariate plot is generated from a labelled dataset
}
\examples{
library(dplyr)
library(ggplot2)
set.seed(1)
spectrum = data.frame(
  Nx = rbinom(50, 100, 0.3), DPx = rep(100, 50),
  Ny = rbinom(50, 100, 0.1), DPy = rep(100, 50),
  node = sample(c("-","+"), 50, replace = TRUE)
)
plot_multivariate(spectrum)
}
