% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_multivariate}
\alias{plot_multivariate}
\title{plotting functions
Plot multivariate VAF distributions with cluster associated to tree nodes.}
\usage{
plot_multivariate(spectrum)
}
\arguments{
\item{spectrum}{VAF spectrum with cluster labels}
}
\value{
A multivariate plot
}
\description{
A multivariate plot is generated from a labelled dataset
}
\examples{
plot_multivariate(spectrum)
}
