% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_pairs}
\alias{plot_pairs}
\title{Pairs plot with divergences marked.}
\usage{
plot_pairs(x, params = NULL)
}
\arguments{
\item{x}{PEPI_Multirates object fit with \code{method = "sample"}.}

\item{params}{Vector of Stan parameter base names. Defaults to a
handful of global scalar parameters (pairs plots grow quadratically
with the number of parameters).}
}
\value{
A ggplot/grid object (see \code{bayesplot::mcmc_pairs()}).
}
\description{
Useful to spot the funnel-shaped regions of parameter space that cause
divergences.
}
\examples{
\dontrun{
plot_pairs(x, params = c("lambda_n","s_epi"))
}
}
